+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cooldown / Throttling support to prevent accidental vibration spam
 - Debug Mode with consistent debug logs and support debug output
 
+### Updated
+
+- Refactor Haptics service
+
 ## [1.0.2] - 2026-05-11
 
 ### Fixed